deepen: inline why-it-matters guard into Summarize module
`lib/ai/parsers/why-it-matters.ts` exported `coerceWhyItMatters` and
`isRestatement` plus a `RestatementCheckInput` interface — 150 LOC of
pure tokenization helpers (stopword set, filing-type-word set, content-
word extraction, consecutive-window overlap, headline-overlap ratio)
behind a 2-function interface.

The module had one production caller: `coerceWhyItMatters` is invoked
once in `lib/ai/summarize.ts` at the post-parse grounding step (around
line 1245). `isRestatement` had zero non-test callers.

Deletion test: removing the module concentrates 150 LOC inside
`lib/ai/summarize.ts` as private functions, but does not reappear at
any other call site — there are none. The interface was the
implementation pretending to be reusable. Same shape as ADR-0002
(`analysis-depth` inlining): a pure function extracted purely for
testability, whose real behavioural coverage lives in the integration
tests at the Summarize seam.

The interface tests in `__tests__/lib/ai/summarize-grounding-wireup.test.ts`
already exercise the guard through `summarizeFiling`: the cases
"drops whyItMatters when it restates the headline (why-it-matters
validator runs)", "preserves grounded whyItMatters when no
contamination or restatement", and "honors DISABLE_WHY_IT_MATTERS_GROUNDING=1
kill switch" cover the production contract. The unit tests in
`why-it-matters.test.ts` (16 cases) tested implementation details past
the interface (specific stopword behavior, 40-char floor, ticker
filtering on the helper functions). Deleted per the seam discipline:
the interface is the test surface.

CONTEXT.md gains a "Why It Matters Guard" entry that pairs with the
existing "Historical Context" entry — both are private helpers inside
the Summarize module that the autonomous architecture review will
otherwise re-suggest extracting.
